+++ Animation/Authoring/BakerExtensions.cs
@@ -3,7 +3,7 @@ using Unity.Entities;
 using Unity.Mathematics;
 using UnityEngine;
 
-namespace NSprites
+namespace NSprites.Authoring
 {
     public static partial class BakerExtensions
     {
