+++ Base/Authoring/Editor/SpriteRendererAuthoringEditor.cs
@@ -1,4 +1,5 @@
-﻿using UnityEditor;
+﻿using NSprites.Authoring;
+using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine;
 using UnityEngine.UIElements;
