+++ Base/Authoring/SpriteRendererAuthoring.cs
@@ -1,4 +1,4 @@
-﻿using NSprites.Modules;
+﻿using NSprites.Authoring;
 using Unity.Entities;
 using Unity.Mathematics;
 using UnityEngine;
